In frame Q adjustment experiment.
This is an early  WIP and not ready for review.

The idea her is to allow "in frame" adjustment of the final Q
value used to encode each SB64, using segmentation.

This could allow frames to more accurately hit a rate target
without the need for recodes, or be used to alter the distribution
of bits between SBs of differing complexity.

Change-Id: Ia9e486e37b3d6cd16e9d65791ff2edb1f71fc300