In frame Q adjustment experiment.
This is a WIP and not ready for review.

The idea her is to allow "in frame" adjustment of the final Q
value used to encode each SB64, using segmentation.

This could allow frames to more accurately hit a rate target
without the need for recodes, or be used to alter the distribution
of bits between SBs of differing complexity.

Change-Id: I2f140cd898c9f877c32cd6d2e667f5e11ada4b1c