Update patch set 1
Patch Set 1: Looks good to me, approved

this is largely cosmetic, but a change in the right direction.
as a follow-up the scope of some of these functions can be reduced, dropping pbi passing if possible. note there are other issues with tiles, that I've had mostly working for some time, so I'll push some of that, though there are lingering tile boundary issues.

Patch-set: 1
Label: Code-Review=+2
Label: Commit-Queue=0