Prevent invalid memory access for Clang builds
Suspecting it is due to an aliasing bug in Clang compiler, a corrupted
input file may lead clang-built decoders to use motion vectors pointing
to area outside of image with mv border, resulting in invalid memory
access. This commit switched the order of mv clamping and assignment
operations to work around the compiler issue.

Change-Id: I24502ff18bd4f9e62507a879cc8760a91a0fd07e