Update patch set 6
Patch Set 6:

Yup, we can merge this once the whitespace in configure.sh is fixed. You can see it in the web interface - trailing spaces on two of the blank lines and some on the last fi.

After that I need to give it a +2 and then we typically wait for Jenkins to +1/Verified as a sanity check, then it can merge.

Patch-set: 6