Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I was not able to build this with Visual Studio 2008 Express or Professional (vs9). The configure script correctly runs and disables AVX/AVX2, but the generated solution will not load the projects within the solution. I get a generic error that the project targets a platform for which support is not installed. I'll look at this again tomorrow and try to determine a root cause for the issue. Note that this happened on x86-win32-vs9 and x86_64-win64-vs9 configure targets.

Patch-set: 3
Reviewer: Gerrit User 1002622 <1002622@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Verified=0