Fix decoder to handle display size correctly
The decoder ignored the display width & height
specified in the frame header.

This patch adds two controls to allow the application
to get the display width and height and modifies
vpxdec to scale the output frame to this size.
(VP8D_GET_DISPLAY_WIDTH and VP8D_GET_DISPLAY_HEIGHT)

Should the request for display size fail vpxdec will
use the native width and height of the raw decoded
frame instead.

Change-Id: I25db04407426dac730263720c75a7dd6400af68a