Update patch set 3
Patch Set 3: Verified

This is more along the lines of where I wanted to start... shame on me for not marking a patchset in RFC state 'DO NOT SUBMIT'. 

Anyway, as it is this setup still isn't going to scale well, but it's a saner starting point than what's been submitted.

Comments welcome!

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=+1