Fix bug in extend_frame chroma extended too far
In the case where the frame was an odd number of pixels
wide or high, the border was being extended by one col
or row too far.

I modified the calculation of color plane dimensions
to use those already computed at the time the frame
buffer was allocated.

Also, I free the temporary scaling buffer to prevent a
memory leak.

Change-Id: Ied04bdcdfd77469731408c05da205db1a6f89bf5