Update patch set 5
Patch Set 5:

The current patch 5 is the merged patch, and it can be downloaded using one of the download methods shown above. For example, the command line for "checkout" is:
git fetch ssh://user@gerrit.chromium.org:29418/webm/libvpx refs/changes/04/67904/5 && git checkout FETCH_HEAD

Yes, the style errors need to be fixed, and committed to get a new patch, and then submit the patch for further review.

Patch-set: 5