Enforce maxq/minq for all frames
This change will always enforce maxq/minq for all types of frames.
The main goal of this patch is to make the maxq/minq parameters
easily explainable and predictable - which will reduce the
confusion many people have.

Besides, the quality boost for altref frames this was intended
for, can still be partially achieved by reducing the minq parameter,
or changing the boost to qp functions. This patch will also
likely reduce the size of alt-ref frames in certain cases where
the minq parameter is greater than 0.

Psnr stats change a small amount for derf/stdhd.

Change-Id: I68ca70448f10864fbbb487a2d2daa82a45040781