Enforce maxq/minq for all frames
This change always enforces maxq/minq for all types of frames.
This makes the maxq/minq parameters much easier to explain.

Besides, the quality boost for altref frames this was intended
for can still be mostly achieved by reducing the minq parameter,
or changing the boost to qp functions.

Psnr stats change a small amount for derf/stdhd.

Change-Id: I68ca70448f10864fbbb487a2d2daa82a45040781