Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The previous code already selected max Q that was always within the user selected min and max Q range. This change is a bit of a clean up but I think the description is misleading

The intent of the original change was to prevent "flattening of Q" when coding at close to a limiting bandwidth for a given user MaxQ setting. Hence even if normal P frames are being coded at MAXQ it is desirable (assuming the range of min and maxQ allows) to apply some delta for alt refs and key frames.

The standard test set does not show the benefit of this change very strongly but I found test cases (hard clips) where it gives a big gain in metrics and visual quality. The size of the delta given here is still up for debate though and may cause increased popping in some cases.

Patch-set: 2
Label: Code-Review=-1
Label: Commit-Queue=0