Update patch set 3
Patch Set 3:

Decoder test using the clip in issue 689 showed that the corruption happens since one motion vector equals MV_LOW. This patch prohibits that.
Verifying if this patch solves issue 689 will take long time because the test clip is a 4-minute-long 4k video, which has to be re-encoded.

Patch-set: 3