Layer based rate control for CBR mode.
This patch adds a buffer-based rate control for layers, for CBR mode.
The current implementation is for temporal layers.

Added an encoder-app vpx_temporal_scalable_patters.c for testing temporal
layers for both vp9 and vp8 (this replaces the old vp8_scalable_patterns).

Added unittests for rate-targeting for 2 and 3 temporal layers.

Change-Id: I9cb6cce2494390ae6096ee17774af7fb9308bde7