Enables alt-ref frames in one pass mode
Includes a few fixes and clean-ups that adds the ability
to use alt-ref frames in one-pass mode. This patch just
adds the ability to turn it on, but whether alt-refs are
actually used or not is controlled by a macro
USE_ALTREF_FOR_ONE_PASS. It is turned off by default.
It will be turned on once we optimize for coding efficiency.

Change-Id: I78254421435478003367c788c7930d2dc4ee2816