Disable update of last_show_frame for existing frame
When showing a previously decoded frame, i.e. when
show_existing_frame=1, the update of the
last_show_frame flag must be disabled.

This is to ensure that last_show_frame flagg reflects
the state of the flag for the immediately previously
decoded frame rather then the value that was forced
to ensure that a previously decoded frame could be
displayed.

(Bug originally reported by Alexander Voronov
<ru.xalba@gmail.com>).

Change-Id: I731d288fba02088959f7fcc87707137fffc6acf5