Update patch set 2
Patch Set 2: Looks good to me, approved

For bilinear, I remember it is possible and faster to do both directions in single pass through as it saves writing to/reading from intermediate results. Of course that does not have to be in this commit, and can be a future optimization.

Patch-set: 2
Label: Code-Review=+2
Label: Commit-Queue=0