Update patch set 1
Patch Set 1: Looks good to me, approved

so this checks if a CFLAG is invalid according to CXX? such as -Wnot_a_real_warning? Does this get around the issue with clang not actually erroring for that one cflag we had an issue with that I can't think of right now?

Patch-set: 1
Label: Code-Review=+2
Label: Commit-Queue=0