Update patch set 6
Patch Set 6: Do not submit

If the largest partition that can fit is 32x16 because we are on the last row and 32x32 doesn't fit ( because say there are only 12 pixels in the last row not 32) this code won't provide a square option.    

I think it should always allow a square partition that is less than the size of the current,  so 0, -3 would work,  though I still think its better to just provide the square partition one down via lookup.

Patch-set: 6
Label: Code-Review=-2