Resolve some VS compiler warnings.
Many of these relate to signed unsigned mismatch for
example comparing to an int loop variable.

In each case int can easily hold the range of the value.
For now several of the structures have been changed to int
but in future we may want to do a broader pass through
to consider where we want to explicitly use unsigned.

Change-Id: Ib1f2ab3640e730dda6895ba29e2d3f7ed9ec6a40