Silence further V.S. warnings.
Note that the use of int for base_q in vp9_vaq.c
seems to me to have been a bug.

There are a number of changes relating to the use
of size_t which can be defined differently on different
platforms. In some cases there are assignments to/from
int or unsigned int. I think the cases in question are safe
but the the inconsistent use of size_t, int and unsigned
int in some cases is a matter of concern.

These changes should be run to make sure they do not
alter results before being submitted.

Apart from AQ mode 1 I do not think they SHOULD alter output.

Change-Id: I470178dd073a2abde4aa98084f16e53370fc8456