Update patch set 1
Patch Set 1:

My main point is that started as a simple example that does a good job of demonstrating only the encoder. Now a third of the code is devoted to parsing the command line: All of the parsing related noise is not useful for people wanting to learn to use the codec.

Patch-set: 1