Update patch set 1
Patch Set 1: (2 inline comments)

This CL is a pure refactor (I'll have a regression test demonstrating that shortly). If it highlights decisions that we are making poorly, that's a benefit of the cleaner code. Given the size of this refactor, new heuristics should probably follow in a fresh CL.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0