Update patch set 2
Patch Set 2: Do not submit

This is required until all the code is moved to intrinsics. It pushes callee-saved registers so the individual functions don't need to.

It actually should be safe to remove now, since I believe most of the vp8 assembly manages these registers correctly, but I'm not 100% sure and it's somewhat difficult to test.

You can drop this patch or move it to the end of the vp8 queue.

Patch-set: 2
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0