Adds variance based fixed size partitioning
Adds a method for determining a fixed size partition based on
variance of a 64x64 SB. This method is added to rtc speed 6.
Also fixes a bug in set_partitioning() and includes some
refactoring related to partitioning search, and some cosmetics.

Currently compared to speed 5, the coding efficiency of speed 6
is -15.8% and that of speed 7 is -55.6% in cbr mode.

Change-Id: I057e04125a8b765906bb7d4bf7a36d1e575de7c6