Motion Vector Test Vector Generator
(DO NOT MERGE WITH CODEBASE -- COMMITTED ONLY
TO DOCUMENT HOW TEST VECTOR WAS GENERATED).

This code generates a test vector to test the
motion vector handling when using scaling.

The code superimposes a 3 frame sequence into the
buffer to be encoded:

Frame 0: CIF Keyframe, completely black.
Frame 1: QCIF inter frame, black with 1 pixel
         thick white border.
Frame 2: CIF inter frame, completely black.

It also forces Q=0.

All blocks in the inter frames are forced to be
coded as NEWMV with a MV defined by which quadrant
the block lies. The MVs are setup to force the
matching block to be placed at the first whole pel
position outside the frame where going further out
would not change the prediction error.

Change-Id: I996ea12628b8e0fdf99432fc7b9f7c196801386c