Further speed trade off adjustments.
Quality is generally a little up for speed 2.
Speed 3 was similar to speed 4 but now positioned more
evenly between 2 and 4 speed and quality wise.
   (derf +4.535% opsnr, yt +7.762% opsnr - even more in ssim)
Speed 4 is a little slower than before but for most clips
much better quality.
   (derf +1.608% opsnr, yt +5.24% opsnr - even more in ssim)
Speed 5 is also a little slower but better quality.

The code has been cleaned up a bit so that for each incremental
speed step changes over the previous speed step are applied.
This makes it easier to see what is changing from one setting to
the next.

Change-Id: I2d98d0d6230af23486adaec01908f58942a7cdeb