Speed 2 alternative balance.
A previous path improved speed 2 quality a little but
more extensive testing showed that it slowed encode
by a few %.

The change will have a similar effect for speed 3 but
should not impact speeds 4+;

This experiment should reverse that and give a speed
up at the cost of a small quality loss.

Borg results pending.

Change-Id: I4493fc1541aaf44587f1a41ff219f7088da9252c