Make last frame source available in current frame encoding
While encoding a frame, its last frame source can be used to give
acurate motion information. This patch prevents last frame to be
overwritten so that it is available during current frame encoding.

To avoid to allocate an extra buffer, the last source is not scaled
in this patch. Compare cpi->Source->y_width/y_height with
cpi->unscaled_last_source->y_width/y_height to see if scaling is
needed.

Change-Id: I0e1ef5e9e1d2badf9d0c7a1a44a7ed5b24c09425