Fix prev_mi assignment issue that caused rt encoding broken
The encoder should be able to use previous frames' coding information
for fast coding mode decision, even when the error resilient mode is
turned on, where such information is not used in the bit-stream coding
stage. This commit fixes an issue where the prev_mi is set to 0x00
at the encoder when error resilient mode is on, which makes the
encoding process broken.

Change-Id: Ica245396bf33be8a5f6e49aeab499abb967d0912