Rate control change to track target closer
This patch turns off the DISABLE_RC_LONG_TERM_MEM macro and
gradually adjusts the bits left from the actual bits used to
the target bits. As a result, the actual bitrate tracks the
target bitrate closer, but the last few frames are not as
starved for bits as the case when the bits left adjustment is
solely based on the actual bits used.

Also changes the inter_minq parameter back to 0.75 from 0.55,
reverting an earlier patch.

Overall derf improves by 0.762%, where the bulk of the gain
comes from the the reverting change.

Change-Id: I0cd1f7edbc518b0b566ab9df0ac34ad2d2360d20