Rate control change to track target closer
This patch turns off the DISABLE_RC_LONG_TERM_MEM macro and
gradually adjusts the bits left from the actual bits used to
the target bits. As a result, the actual bitrate tracks the
target bitrate closer, but the last few frames are not as
starved for bits as the case when the bits left adjustment is
solely based on the actual bits used.

derf actually improves by +0.128% in speed 0, and the bitrate
achieved is closer to the target.

Change-Id: I0cd1f7edbc518b0b566ab9df0ac34ad2d2360d20