Update patch set 2
Patch Set 2:

I was thinking a possible alternate approach might be that we have two frame target numbers. The original bit allocation for each frame (as if DISABLE_RC_LONG_TERM_MEM was set) but a modified value that is used as a target for the frame.

A "bits off target" value is updated based on whether we use more or less bits than the "no memory" value. The modified target is adjusted each frame according to whether the "bits off target" is + or - but only up to a limit of e.g. +/- X% from the no memory value.

This method would correct early (up to the % limit) but would never suffer from extreme bias as no frame can be offered more than +/- X% of its no memory target.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0