Added initial support for 16-bit framebuffers.
Changes in this patch are only enabled if configured with
--enable-experimental --enable-vp9_high

Using a encoder command line argument of --input-shift=0 tells the coder
to work with 16bit framebuffers.
The output should be identical to before. Some features (such as input
image resizing) are not yet supported in 16bit mode.

Specifically, the behavior of the input-shift parameter is as follows:
* No argument : Behaviour as before, using 8bit frame buffers
* --experimental-bitstream --profile=2 --input-shift=0: Uses
  16bit frame buffers to store 8-bit data, should give identical output
  to before.
* --experimental-bitstream --profile=2 --input-shift=2 --bit-depth=1: Uses
  16bit frame buffers to store 10-bit data, encodes a version 2 stream
  with bitdepth 10
* --experimental-bitstream --profile=2 --input-shift=4 --bit-depth=2: Uses
  16bit frame buffers to store 12-bit data, encodes a version 2 stream
  with bitdepth 12

The decoder has a --output-shift argument which should be used when
decoding profile 2 streams.

So far support for the following has been added:
Intra filtering
Deblocking
Motion compensation
Variance calculation
Sad calculation
Transform

Change-Id: If345c88234aafdd40caea0d88935b1f07aaebe22