Remove ARNR filter type from code completely
The ARNR filter type was originally added for the
purpose of experimentation, but only the centered
filter is considered to be useful.

Consequently, this patch removes the option from
the code and hard-codes the use of the centered
option. A previous patch removed this option from
VP9.

The command line option has been removed from vpxenc
as part of this change.

Note: This change will require a corresponding
change in ffmpeg and all existing script files will
need to be modified accordingly.

Change-Id: Ie431448128b4aeba3c78f7fcb341533d4684550f