Update patch set 14
Patch Set 14:

@Pieter

I share your concerns and there may well be a better way to write the code that highlights the commonality.

Part of the problem is that there are a lot of hardcoded assembler routines that implement the current 8-bit functions so it is not desirable to change the 8-bit function prototypes.

However, note that the new functions working on 16bit framebuffers form a complete implementation of all bitdepths.  In particular, the new functions support 8bit depth in a way that can be tested to be identical to the old functions.

My hope is that if you look just at the new functions you will see that they do indeed take an extra parameter specifying the bitdepth as you suggest (where necessary) and so it should be easy to see the effect of high bitdepth.

(By the way, this is just a partial code drop and does not reflect the full scope of the proposed changes - in particular there will be more changes to transform and residual coding to increase the range.)

Patch-set: 14
Reviewer: Gerrit User 1115299 <1115299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0