Circumvents : Possible win64 msvc compile issue?
This proposed fix seems to circumvent.

https://code.google.com/p/chromium/issues/detail?id=370882

Some quick attempts at duplicating the bug in a very simple console app
failed.  Apparently with optimize size ( -O1 and -Os ) the compiler
attempts to convert the pointer to the union struct ( 2 shorts unioned
with an int), into the actual value for that int and then attempts to
de reference it.

TODO(jbb) : Try harder to fix this and either file a bug against
msvc or find a bug that seems likely to be.   Also watch for future
versions to see if this is a problem.

Change-Id: I03339f631f507cc008759e2fae4321bc10d30b91