Circumvents : Possible win64 msvc compile issue?
This proposed fix seems to circumvent.

https://code.google.com/p/chromium/issues/detail?id=370882

Some quick attempts at duplicating the bug in a very simple console app
failed.  Apparently with optimize size ( -O1 and -Os ) the compiler
attempts to convert the pointer to the union struct ( 2 shorts unioned
with an int), into the a pass by value and then when passed to a
subfunction it attempts to dereference it and causes a memory allocation
problem.

TODO(jbb) : Try harder to duplicate this in a sample app and then either
file a bug against msvc or find a similar already posted bug. In either
case check future msvc builds to see if the problem persists.

Change-Id: I03339f631f507cc008759e2fae4321bc10d30b91