Update patch set 2
Patch Set 2:

The 0.9 value is very tight and gives close to fixed Q for P frames. It may well be this is just to aggressive even if in the 2 pass it is helping a little on metrics. Also, for RTC we really don't have regular alt refs etc to spend the bits on, so I think you are right that we must allow a bigger spread. Did you test even lower?

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0