Adding new graphical bitstream analysis tool, WIP.
QT5 application for bitstream analysis. Right now the UI is structurally complete, but no functionality.
The Bitstream class in bitstream.h is the wrapper for holding the libvpx vp9 decoder.
Data to be analyzed is captured with protocol buffers as detailed in vp9bat.proto.
This makes it easy for a future project to re-use the non QT code to create for example an HTML5 front end.

Running this requires have protobuf library installed, and QT5 installed. It can be compiled with QT Creator, by opening the .pro file, or by running qmake on it, followed by the resulting makefile.
Eventually this all could be intergrated in the main build flow of libvpx.

Properly storing all information from the decode process will require some changes to the libvpx vp9 decoder rountines.
My intention is to add #ifdefs for these edits.

Change-Id: I9a1d31821cac8be89df0c4a5ca44e56748b61b15