LONG_TERM_VBR_CORRECTION reinstated.
More detailed analysis shows that the alternative
method is in some situations giving extremely unstable
rate recommendations frame to frame. The impact
of this has been masked by the use of a narrow
range of allowed Q values for most normal frames
(active min and max) but it is still a cause of concern.

On our test sets this change is slightly negative in terms
of metrics, but I suspect this is just noise relating to how
aggressively the rate control tries to hit its target.

Change-Id: Ic90ed36af129c006dfc81869e09bf7713b402a55