Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

For a non-super frame, there will be only one frame in the trunk. Also, decode_one api is still the same that accepting pointer to a pointer which is bad for frame parallel. So even this patch is in, I still have another one that will override this to make it work.

Patch-set: 2