Refactor decoder_decode for superframe.
A superframe is a bunch of frames that bundled as one frame. It is mostly
used to combine one or more non-displayable frames and one displayable frame.

If a frame that passed to decoder_decode is not a superframe, there
should be only one displayable frame in it. If a frame that passed to
decoder_decode is a superframe, then superframe index should be presented
with the frame. Then each frame's size could be retrieved from that.

Change-Id: I06a5716c8608c5e3585b776601e90fba587fe329