Correct the unit test for superframe.
A superframe is a bunch of frames that bundled as one frame. It is mostly
used to combine one or more non-displayable frames and one displayable frame.

If a frame that passed to decoder_decode is not a superframe, there
should be only one displayable frame in it. If a frame that passed to
decoder_decode is a superframe, then superframe index should be presented
with the frame. Then each frame's size could be retrieved from that.

The original test try to get rid of the superframe index and send a chunk
of frames into the decoder_decode which should not be supported by the decoder.

Change-Id: If9a29ecb290c3c9cee66c092c06f4356714d84e1